[CLEANUP] Fixed most of the issues that caused compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@1095 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ models/ApiFunction.cpp
@@ -335,7 +335,7 @@ void ApiFunction::generateToolTipText(unsigned int paramIndex, QString& text) co
 
     text += ")<br>" + desc_;
     
-    if (paramIndex < params_.size())
+    if (paramIndex < static_cast<unsigned int>(params_.size()))
     {
         text += "<br><b>Parameter:</b> " + params_.at(paramIndex)->getDescription();
     }
